Fine-map locus with FINEMAP

The stepwise conditional search starts with a causal configuration containing the SNP with the lowest P-value alone and then iteratively adds to the causal configuration the SNP given the highest posterior model probability until no further SNP yields a higher posterior model probability.

Output coumns
Note that not not all versions of FINEMAP will necessarily have all of these columns, but "PP" and "CS" will always be present regardless of version.

PP

Per-SNP conditional posterior probability (PP), after conditioning on the other SNPs within its respective Credible Set (CS). If a given SNP is present in more than one CS, each row will be a list of PP (one per CS). These PP were extracted from .cred file(s). SNPs that were not within the CS are designated NA.

CS

The Credible Set (CS) that a given SNP belongs to. If a given SNP is present in more than one CS, each row will be a list of CS id numbers (one per CS). These CS were extracted from .cred file(s).

PP_snp

Per-SNP marginal posterior inclusion probability (PIP), which is the probability that a given SNP is in the Credible Set (CS). These PIP were extracted from .snp file(s). Only SNPs that were excluded from the fine-mapping input (e.g. due to not overlapping with the LD panel) are designated NA.

PP_config

Per-CS posterior probability (PP) that a given Credible Set (CS) (i.e. "configuration" of SNPs) is causal. These PP were extracted from .config file(s). Only SNPs that were not within any CS are designated NA.

k

Optimized number of causal SNPs, which can be less than or equal to the user-supplied n_causal argument. These values were extracted from .log file(s).

FINEMAP(
  dat,
  locus_dir = tempdir(),
  LD_matrix,
  FINEMAP_path = NULL,
  compute_n = "ldsc",
  n_causal = 5,
  model = c("sss", "cond"),
  remove_tmps = FALSE,
  force_new = FALSE,
  credset_thresh = 0.95,
  finemap_version = package_version("1.4.1"),
  priors_col = NULL,
  rescale_priors = TRUE,
  args_list = list(),
  fillNA = 0,
  nThread = 1,
  verbose = TRUE
)

Source

http://www.christianbenner.com

Arguments

dat

Fine-mapping results data.

locus_dir

Locus-specific directory to store results in.

LD_matrix

Linkage Disequilibrium (LD) matrix to use for fine-mapping.

FINEMAP_path

Path to a custom FINEMAP executable to use instead of the ones included in echolocatoR. Users can also simply supply "finemap" if this command is linked to the executable.

compute_n

How to compute per-SNP sample size (new column "N").
If the column "N" is already present in dat, this column will be used to extract per-SNP sample sizes and the argument compute_n will be ignored.
If the column "N" is not present in dat, one of the following options can be supplied to compute_n:

0

N will not be computed.

>0

If any number >0 is provided, that value will be set as N for every row. **Note**: Computing N this way is incorrect and should be avoided if at all possible.

"sum"

N will be computed as: cases (N_CAS) + controls (N_CON), so long as both columns are present.

"ldsc"

N will be computed as effective sample size: Neff =(N_CAS+N_CON)*(N_CAS/(N_CAS+N_CON)) / mean((N_CAS/(N_CAS+N_CON))(N_CAS+N_CON)==max(N_CAS+N_CON)).

"giant"

N will be computed as effective sample size: Neff = 2 / (1/N_CAS + 1/N_CON).

"metal"

N will be computed as effective sample size: Neff = 4 / (1/N_CAS + 1/N_CON).

n_causal

The maximum number of potential causal SNPs per locus. This parameter is used somewhat differently by different fine-mapping tools. See tool-specific functions for details.

model

"cond" for stepwise conditional search, "sss" for stochastic shotgun search.

remove_tmps

Remove any temporary files generated.

force_new

If saved results already exist in the given locus_dir, skip re-running FINEMAP and use them (default: force_new). Set TRUE to ignore these files and re-run FINEMAP.

credset_thresh

The minimum mean Posterior Probability (across all fine-mapping methods used) of SNPs to be included in the "mean.CS" column.

finemap_version

Which FINEMAP version to use (specify as a string).

priors_col

[Optional] Name of the a column in dat to extract SNP-wise prior probabilities from.

rescale_priors

If prior probabilities are supplied, rescale them from 0-1 (i.e. rescaled_priors = priors / sum(priors)).

args_list

A named list of additional arguments to pass to FINEMAP (e.g.: args_list = list("–n-iterations"=5000,"–sss"="")). Alternatively, can supply a string instead (e.g.: args_list = "–n-iterations 5000 –sss").

fillNA

Fill CS/PP values without fine-mapping results (i.e. NA) with some default value (e.g. 0).

nThread

Number of threads to parallelise across. Passed to "--n-threads" in FINEMAP.

verbose

Print messages.

See also

Examples

locus_dir <- file.path(tempdir(),echodata::locus_dir)
dat <- echodata::BST1;
LD_matrix <- echofinemap::drop_finemap_cols(echodata::BST1_LD_matrix)
#> Gathering method sources.
#> Gathering method citations.
out <- echoLD::subset_common_snps(LD_matrix, dat)
LD_matrix <- out$LD
dat <- out$DT

dat2 <- echofinemap::FINEMAP(dat=dat,
                             locus_dir=locus_dir,
                             LD_matrix=LD_matrix) 
#> FINEMAP path: /github/home/.cache/R/echofinemap/FINEMAP/finemap_v1.4.1_x86_64/finemap_v1.4.1_x86_64
#> Checking FINEMAP is runnable...
#> FINEMAP is runnable.
#> Inferred FINEMAP version: 1.4.1
#> Loading required namespace: MungeSumstats
#> Preparing sample size column (N).
#> Computing effective sample size using the LDSC method:
#>  Neff = (N_CAS+N_CON) * (N_CAS/(N_CAS+N_CON)) / mean((N_CAS/(N_CAS+N_CON))[(N_CAS+N_CON)==max(N_CAS+N_CON)]))
#> + Mapping colnames from MungeSumstats ==> echolocatoR
#> + Subsetting LD matrix and dat to common SNPs...
#> Removing unnamed rows/cols
#> Replacing NAs with 0
#> + LD_matrix = 95 SNPs.
#> + dat = 95 SNPs.
#> + 95 SNPs in common.
#> Converting obj to sparseMatrix.
#> Constructing master file.
#> Constructing data.z file.
#> Constructing data.ld file.
#> Running FINEMAP.
#> cd .../BST1 &&
#>     .../finemap_v1.4.1_x86_64
#>    
#>     --sss
#>    
#>     --in-files .../master
#>    
#>     --log
#>    
#>     --n-causal-snps 5
#> 
#> |--------------------------------------|
#> | Welcome to FINEMAP v1.4.1            |
#> |                                      |
#> | (c) 2015-2022 University of Helsinki |
#> |                                      |
#> | Help :                               |
#> | - ./finemap --help                   |
#> | - www.finemap.me                     |
#> | - www.christianbenner.com            |
#> |                                      |
#> | Contact :                            |
#> | - finemap@christianbenner.com        |
#> | - matti.pirinen@helsinki.fi          |
#> |--------------------------------------|
#> 
#> --------
#> SETTINGS
#> --------
#> - dataset            : all
#> - corr-config        : 0.95
#> - n-causal-snps      : 5
#> - n-configs-top      : 50000
#> - n-conv-sss         : 100
#> - n-iter             : 100000
#> - n-threads          : 1
#> - prior-k0           : 0
#> - prior-std          : 0.05 
#> - prob-conv-sss-tol  : 0.001
#> - prob-cred-set      : 0.95
#> 
#> ------------
#> FINE-MAPPING (1/1)
#> ------------
#> - GWAS summary stats               : FINEMAP/data.z
#> - SNP correlations                 : FINEMAP/data.ld
#> - Causal SNP stats                 : FINEMAP/data.snp
#> - Causal configurations            : FINEMAP/data.config
#> - Credible sets                    : FINEMAP/data.cred
#> - Log file                         : FINEMAP/data.log_sss
#> - Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary statistics       : -- Reading summary statistics       : +- Reading summary stats            : done!   - Reading SNP correlations         : 0.000%- Reading SNP correlations         : done!   - Reading input                    : done!   
#> 
#> - Number of GWAS samples           : 1474097
#> - Number of SNPs                   : 95
#> - Prior-Pr(# of causal SNPs is k)  : 
#>   (0 -> 0)
#>    1 -> 0.584
#>    2 -> 0.292
#>    3 -> 0.0963
#>    4 -> 0.0236
#>    5 -> 0.00456
#> 0 configurations evaluated (0.000/100%)- 95 configurations evaluated (0.000/100%)- 128 configurations evaluated (0.000/100%)- 2027 configurations evaluated (0.050/100%)- 2027 configurations evaluated (0.100/100%)- 2027 configurations evaluated (0.105/100%) : converged after 105 iterations
#> - Computing causal SNP statistics  : 0.049%- Computing causal SNP statistics  : 2.516%- Computing causal SNP statistics  : 4.983%- Computing causal SNP statistics  : 7.449%- Computing causal SNP statistics  : 9.916%- Computing causal SNP statistics  : 12.383%- Computing causal SNP statistics  : 14.850%- Computing causal SNP statistics  : 17.316%- Computing causal SNP statistics  : 19.783%- Computing causal SNP statistics  : 22.250%- Computing causal SNP statistics  : 24.716%- Computing causal SNP statistics  : 27.183%- Computing causal SNP statistics  : 29.650%- Computing causal SNP statistics  : 32.116%- Computing causal SNP statistics  : 34.583%- Computing causal SNP statistics  : 37.050%- Computing causal SNP statistics  : 39.517%- Computing causal SNP statistics  : 41.983%- Computing causal SNP statistics  : 44.450%- Computing causal SNP statistics  : 46.917%- Computing causal SNP statistics  : 49.383%- Computing causal SNP statistics  : 51.850%- Computing causal SNP statistics  : 54.317%- Computing causal SNP statistics  : 56.783%- Computing causal SNP statistics  : 59.250%- Computing causal SNP statistics  : 61.717%- Computing causal SNP statistics  : 64.184%- Computing causal SNP statistics  : 66.650%- Computing causal SNP statistics  : 69.117%- Computing causal SNP statistics  : 71.584%- Computing causal SNP statistics  : 74.050%- Computing causal SNP statistics  : 76.517%- Computing causal SNP statistics  : 78.984%- Computing causal SNP statistics  : 81.450%- Computing causal SNP statistics  : 83.917%- Computing causal SNP statistics  : 86.384%- Computing causal SNP statistics  : 88.851%- Computing causal SNP statistics  : 91.317%- Computing causal SNP statistics  : 93.784%- Computing causal SNP statistics  : 96.251%- Computing causal SNP statistics  : 98.717%- Computing causal SNP statistics  : done!   
#> - Regional SNP heritability        : 0.00296 (SD: 8.93e-05 ; 95% CI: [0.00279,0.00314])
#> - Log10-BF of >= one causal SNP    : 931
#> - Post-expected # of causal SNPs   : 5
#> - Post-Pr(# of causal SNPs is k)   : 
#>   (0 -> 0)
#>    1 -> 0
#>    2 -> 0
#>    3 -> 0
#>    4 -> 2.61e-49
#>    5 -> 1
#> - Computing credible sets          : - Writing causal configurations    : 0.049%- Writing causal configurations    : 2.516%- Writing causal configurations    : 4.983%- Writing causal configurations    : 7.449%- Writing causal configurations    : 9.916%- Writing causal configurations    : 12.383%- Writing causal configurations    : 14.850%- Writing causal configurations    : 17.316%- Writing causal configurations    : 19.783%- Writing causal configurations    : 22.250%- Writing causal configurations    : 24.716%- Writing causal configurations    : 27.183%- Writing causal configurations    : 29.650%- Writing causal configurations    : 32.116%- Writing causal configurations    : 34.583%- Writing causal configurations    : 37.050%- Writing causal configurations    : 39.517%- Writing causal configurations    : 41.983%- Writing causal configurations    : 44.450%- Writing causal configurations    : 46.917%- Writing causal configurations    : 49.383%- Writing causal configurations    : 51.850%- Writing causal configurations    : 54.317%- Writing causal configurations    : 56.783%- Writing causal configurations    : 59.250%- Writing causal configurations    : 61.717%- Writing causal configurations    : 64.184%- Writing causal configurations    : 66.650%- Writing causal configurations    : 69.117%- Writing causal configurations    : 71.584%- Writing causal configurations    : 74.050%- Writing causal configurations    : 76.517%- Writing causal configurations    : 78.984%- Writing causal configurations    : 81.450%- Writing causal configurations    : 83.917%- Writing causal configurations    : 86.384%- Writing causal configurations    : 88.851%- Writing causal configurations    : 91.317%- Writing causal configurations    : 93.784%- Writing causal configurations    : 96.251%- Writing causal configurations    : 98.717%- Writing causal SNP statistics    : 1.053%- Writing causal SNP statistics    : 11.579%- Writing causal SNP statistics    : 22.105%- Writing causal SNP statistics    : 32.632%- Writing causal SNP statistics    : 43.158%- Writing causal SNP statistics    : 53.684%- Writing causal SNP statistics    : 64.211%- Writing causal SNP statistics    : 74.737%- Writing causal SNP statistics    : 85.263%- Writing causal SNP statistics    : 95.789%- Writing output                   : done!   
#> - Run time                         : 0 hours, 0 minutes, 0 seconds
#> 1 data.cred* file(s) found in the same subfolder.
#> Selected file based on postPr_k: data.cred5
#> Importing conditional probabilities (.cred file).
#> Importing marginal probabilities (.snp file).
#> Importing configuration probabilities (.config file).